+++ web/src/layouts/SidebarLayout/SidebarLayout.tsx
@@ -1,220 +1,17 @@
 import { Fragment, useState } from 'react'
 
-import { useQuery } from '@apollo/client'
 import { Dialog, Transition } from '@headlessui/react'
-import { Flex, Subtitle, Title } from '@tremor/react'
-import { GetConnectionStatus } from 'types/graphql'
 
-import { NavLink, routes } from '@redwoodjs/router'
+import { CloseIcon, MenuIcon } from 'src/icons/Icons'
 
-import {
-  CloseIcon,
-  DashboardIcon,
-  DiscordIcon,
-  DiscourseIcon,
-  ErdIcon,
-  GitHubIcon,
-  GraphQLIcon,
-  HomeIcon,
-  InboxIcon,
-  MenuIcon,
-  OGTagPreviewIcon,
-  OperationsIcon,
-  PlaygroundIcon,
-  SpansIcon,
-  SqlStatementsIcon,
-  TemplatesIcon,
-  TracesIcon,
-} from 'src/icons/Icons'
-
-const CONNECTION_STATUS_QUERY = gql`
-  query GetConnectionStatus @live {
-    connectionStatus {
-      id
-      developmentServer
-    }
-  }
-`
+import { NavigationMenu } from './NavigationMenu'
 
 type SidebarLayoutProps = {
   children?: React.ReactNode
 }
 
-type TNavigationItem = {
-  name: string
-  to: string
-  icon: typeof HomeIcon | typeof GraphQLIcon | typeof DiscordIcon
-}
-
-// v7 ships with SSR section disabled as Bighorn introduces the feature
-const includeSSR = false
-
-const NavigationItem = ({ item }: { item: TNavigationItem }) => {
-  return (
-    <div className="flex w-full">
-      <NavLink
-        to={item.to}
-        className="group flex grow gap-x-3 rounded-md p-2 text-sm font-semibold leading-6 text-gray-500 hover:bg-gray-50 hover:text-orange-500 dark:text-gray-600 dark:hover:bg-gray-950"
-        activeClassName="grow text-orange-500 dark:text-orange-500 bg-gray-50 dark:bg-gray-950 group flex gap-x-3 rounded-md p-2 text-sm leading-6 font-semibold"
-      >
-        <item.icon className="h-6 w-6 shrink-0" aria-hidden="true" />
-        {item.name}
-      </NavLink>
-    </div>
-  )
-}
-
-const NavigationHrefItem = ({ item }: { item: TNavigationItem }) => {
-  return (
-    <div className="flex w-full">
-      <a
-        href={item.to}
-        target="_blank"
-        className="group flex grow gap-x-3 rounded-md p-2 text-sm font-semibold leading-6 text-gray-500 hover:bg-gray-50 hover:text-orange-500 dark:text-gray-600 dark:hover:bg-gray-950"
-        rel="noreferrer"
-      >
-        <item.icon className="h-6 w-6 shrink-0" aria-hidden="true" />
-        {item.name}
-      </a>
-    </div>
-  )
-}
-
-const ConnectionStatusIndicator = ({
-  developmentServer,
-}: {
-  developmentServer: 'connected' | 'disconnected' | 'unknown'
-}) => {
-  const bgColorDark =
-    developmentServer === 'connected'
-      ? 'bg-green-500'
-      : developmentServer === 'disconnected'
-      ? 'bg-red-500'
-      : 'bg-yellow-500'
-  const bgColorLight =
-    developmentServer === 'connected'
-      ? 'bg-green-400'
-      : developmentServer === 'disconnected'
-      ? 'bg-red-400'
-      : 'bg-yellow-400'
-
-  return (
-    <div className="group flex w-full grow items-center gap-x-3 rounded-md p-2 text-sm font-semibold leading-6 text-gray-500 dark:text-gray-600">
-      <span className="relative flex h-3 w-3">
-        <span
-          className={`absolute inline-flex h-full w-full animate-ping rounded-full ${bgColorLight} opacity-75`}
-        ></span>
-        <span
-          className={`relative inline-flex h-3 w-3 rounded-full ${bgColorDark}`}
-        ></span>
-      </span>
-      <span>
-        {developmentServer === 'connected'
-          ? 'Connected'
-          : developmentServer === 'disconnected'
-          ? 'Disconnected'
-          : 'Please Reload'}
-      </span>
-    </div>
-  )
-}
-
 const SidebarLayout = ({ children }: SidebarLayoutProps) => {
   const [sidebarOpen, setSidebarOpen] = useState(false)
-
-  const navigation: TNavigationItem[] = [
-    { name: 'Dashboard', to: routes.home(), icon: DashboardIcon },
-  ]
-  const telemetryNavigation: TNavigationItem[] = [
-    {
-      name: 'Traces',
-      to: routes.opentelemetryTraces(),
-      icon: TracesIcon,
-    },
-    {
-      name: 'Spans',
-      to: routes.opentelemetrySpans(),
-      icon: SpansIcon,
-    },
-  ]
-  const databaseNavigation: TNavigationItem[] = [
-    {
-      name: 'Schema',
-      to: routes.databaseSchema(),
-      icon: ErdIcon,
-    },
-    {
-      name: 'SQL Statements',
-      to: routes.databaseSqlStatements(),
-      icon: SqlStatementsIcon,
-    },
-  ]
-
-  const graphQLNavigation: TNavigationItem[] = [
-    {
-      name: 'Playground',
-      to: routes.graphiql(),
-      icon: PlaygroundIcon,
-    },
-    {
-      name: 'Operations',
-      to: routes.graphQLOperations(),
-      icon: OperationsIcon,
-    },
-    {
-      name: 'Schema Diagram',
-      to: routes.graphqlSchema(),
-      icon: GraphQLIcon,
-    },
-  ]
-  const mailerNavigation: TNavigationItem[] = [
-    { name: 'Inbox', to: routes.mailerInbox(), icon: InboxIcon },
-    {
-      name: 'Templates',
-      to: routes.mailerTemplatePreview(),
-      icon: TemplatesIcon,
-    },
-  ]
-  const ssrNavigation: TNavigationItem[] = [
-    {
-      name: 'OG Tag Preview',
-      to: routes.ogTagPreview(),
-      icon: OGTagPreviewIcon,
-    },
-  ]
-
-  const communityNavigation: TNavigationItem[] = [
-    {
-      name: 'Discord',
-      to: 'https://discord.gg/redwoodjs',
-      icon: DiscordIcon,
-    },
-    {
-      name: 'Forums',
-      to: 'https://community.redwoodjs.com/',
-      icon: DiscourseIcon,
-    },
-    {
-      name: 'GitHub',
-      to: 'https://github.com/redwoodjs',
-      icon: GitHubIcon,
-    },
-  ]
-
-  const {
-    data: connectionStatusData,
-    loading: connectionStatusLoading,
-    error: connectionStatusError,
-  } = useQuery<GetConnectionStatus>(CONNECTION_STATUS_QUERY)
-
-  const developmentServerStatus =
-    connectionStatusLoading ||
-    connectionStatusError ||
-    connectionStatusData === undefined
-      ? 'unknown'
-      : connectionStatusData.connectionStatus.developmentServer
-      ? 'connected'
-      : 'disconnected'
 
   return (
     <>
@@ -271,123 +68,8 @@ const SidebarLayout = ({ children }: SidebarLayoutProps) => {
                       </button>
                     </div>
                   </Transition.Child>
-                  {/* Sidebar component, swap this element with another sidebar if you like */}
                   <div className="flex grow flex-col gap-y-5 overflow-y-auto bg-white px-6 pb-2 dark:bg-gray-900">
-                    <div className="flex h-16 shrink-0 items-center">
-                      <img
-                        className="h-8 w-auto"
-                        src="/mark.svg"
-                        alt="RedwoodJS"
-                      />
-                      <Flex
-                        flexDirection="col"
-                        justifyContent="start"
-                        alignItems="start"
-                      >
-                        <Title className="pl-4">RedwoodJS Studio</Title>
-                        <Subtitle className="-mt-1 pl-6 text-sm">
-                          {window.RW_STUDIO_VERSION}
-                        </Subtitle>
-                      </Flex>
-                    </div>
-                    <nav className="flex flex-1 flex-col">
-                      <ul className="flex flex-1 flex-col gap-y-7">
-                        <li>
-                          <ul className="-mx-2 space-y-1">
-                            {navigation.map((item) => (
-                              <li key={item.name}>
-                                <NavigationItem item={item} />
-                              </li>
-                            ))}
-                          </ul>
-                        </li>
-                        <li>
-                          <div className="text-xs font-semibold leading-6 text-gray-400">
-                            Monitoring
-                          </div>
-                          <ul className="-mx-2 mt-2 space-y-1">
-                            {telemetryNavigation.map((item) => (
-                              <li key={item.name}>
-                                <NavigationItem item={item} />
-                              </li>
-                            ))}
-                          </ul>
-                        </li>
-                        <li>
-                          <div className="text-xs font-semibold leading-6 text-gray-400">
-                            Database
-                          </div>
-                          <ul className="-mx-2 mt-2 space-y-1">
-                            {databaseNavigation.map((item) => (
-                              <li key={item.name}>
-                                <NavigationItem item={item} />
-                              </li>
-                            ))}
-                          </ul>
-                        </li>
-                        <li>
-                          <div className="text-xs font-semibold leading-6 text-gray-400">
-                            GraphQL
-                          </div>
-                          <ul className="-mx-2 mt-2 space-y-1">
-                            {graphQLNavigation.map((item) => (
-                              <li key={item.name}>
-                                <NavigationItem item={item} />
-                              </li>
-                            ))}
-                          </ul>
-                        </li>
-                        <li>
-                          <div className="text-xs font-semibold leading-6 text-gray-400">
-                            Mailer
-                          </div>
-                          <ul className="-mx-2 mt-2 space-y-1">
-                            {mailerNavigation.map((item) => (
-                              <li key={item.name}>
-                                <NavigationItem item={item} />
-                              </li>
-                            ))}
-                          </ul>
-                        </li>
-                        <li>
-                          <div className="text-xs font-semibold leading-6 text-gray-400">
-                            Community
-                          </div>
-                          <ul className="-mx-2 mt-2 space-y-1">
-                            {communityNavigation.map((item) => (
-                              <li key={item.name}>
-                                <NavigationHrefItem item={item} />
-                              </li>
-                            ))}
-                          </ul>
-                        </li>
-
-                        {includeSSR && (
-                          <li>
-                            <div className="text-xs font-semibold leading-6 text-gray-400">
-                              SSR
-                            </div>
-                            <ul className="-mx-2 mt-2 space-y-1">
-                              {ssrNavigation.map((item) => (
-                                <li key={item.name}>
-                                  <NavigationItem item={item} />
-                                </li>
-                              ))}
-                            </ul>
-                          </li>
-                        )}
-                        <li className="mb-6 mt-auto">
-                          <div className="text-xs font-semibold leading-6 text-gray-400">
-                            Development Server
-                          </div>
-                          <ul className="-mx-2 mt-2 space-y-1">
-                            <ConnectionStatusIndicator
-                              developmentServer={developmentServerStatus}
-                            />
-                          </ul>
-                        </li>
-                      </ul>
-                    </nav>
+                    <NavigationMenu />
                   </div>
                 </Dialog.Panel>
               </Transition.Child>
@@ -398,117 +80,7 @@ const SidebarLayout = ({ children }: SidebarLayoutProps) => {
         {/* Static sidebar for desktop */}
         <div className="hidden lg:fixed lg:inset-y-0 lg:z-50 lg:flex lg:w-72 lg:flex-col">
           <div className="flex grow flex-col gap-y-5 overflow-y-auto border-r border-gray-200 bg-white px-6 dark:border-gray-800 dark:bg-gray-900">
-            <div className="flex h-16 shrink-0 items-center">
-              <img className="h-8 w-auto" src="/mark.svg" alt="RedwoodJS" />
-              <Flex
-                flexDirection="col"
-                justifyContent="start"
-                alignItems="start"
-              >
-                <Title className="pl-4">RedwoodJS Studio</Title>
-                <Subtitle className="-mt-1 pl-6 text-sm">
-                  {window.RW_STUDIO_VERSION}
-                </Subtitle>
-              </Flex>
-            </div>
-            <nav className="flex flex-1 flex-col">
-              <ul className="flex flex-1 flex-col gap-y-4">
-                <li>
-                  <ul className="-mx-2 space-y-1">
-                    {navigation.map((item) => (
-                      <li key={item.name}>
-                        <NavigationItem item={item} />
-                      </li>
-                    ))}
-                  </ul>
-                </li>
-                <li>
-                  <div className="text-xs font-semibold leading-6 text-gray-400 dark:text-gray-600">
-                    Monitoring
-                  </div>
-                  <ul className="-mx-2 mt-2 space-y-1">
-                    {telemetryNavigation.map((item) => (
-                      <li key={item.name}>
-                        <NavigationItem item={item} />
-                      </li>
-                    ))}
-                  </ul>
-                </li>
-                <li>
-                  <div className="text-xs font-semibold leading-6 text-gray-400 dark:text-gray-600">
-                    GraphQL
-                  </div>
-                  <ul className="-mx-2 mt-2 space-y-1">
-                    {graphQLNavigation.map((item) => (
-                      <li key={item.name}>
-                        <NavigationItem item={item} />
-                      </li>
-                    ))}
-                  </ul>
-                </li>
-                <li>
-                  <div className="text-xs font-semibold leading-6 text-gray-400 dark:text-gray-600">
-                    Database
-                  </div>
-                  <ul className="-mx-2 mt-2 space-y-1">
-                    {databaseNavigation.map((item) => (
-                      <li key={item.name}>
-                        <NavigationItem item={item} />
-                      </li>
-                    ))}
-                  </ul>
-                </li>
-                <li>
-                  <div className="text-xs font-semibold leading-6 text-gray-400 dark:text-gray-600">
-                    Mailer
-                  </div>
-                  <ul className="-mx-2 mt-2 space-y-1">
-                    {mailerNavigation.map((item) => (
-                      <li key={item.name}>
-                        <NavigationItem item={item} />
-                      </li>
-                    ))}
-                  </ul>
-                </li>
-                <li>
-                  <div className="text-xs font-semibold leading-6 text-gray-400 dark:text-gray-600">
-                    Community
-                  </div>
-                  <ul className="-mx-2 mt-2 space-y-1">
-                    {communityNavigation.map((item) => (
-                      <li key={item.name}>
-                        <NavigationHrefItem item={item} />
-                      </li>
-                    ))}
-                  </ul>
-                </li>
-
-                {includeSSR && (
-                  <li>
-                    <div className="text-xs font-semibold leading-6 text-gray-400 dark:text-gray-600">
-                      SSR
-                    </div>
-                    <ul className="-mx-2 mt-2 space-y-1">
-                      {ssrNavigation.map((item) => (
-                        <li key={item.name}>
-                          <NavigationItem item={item} />
-                        </li>
-                      ))}
-                    </ul>
-                  </li>
-                )}
-                <li className="mb-6 mt-auto">
-                  <div className="text-xs font-semibold leading-6 text-gray-400 dark:text-gray-600">
-                    Development Server
-                  </div>
-                  <ul className="-mx-2 mt-2 space-y-1">
-                    <ConnectionStatusIndicator
-                      developmentServer={developmentServerStatus}
-                    />
-                  </ul>
-                </li>
-              </ul>
-            </nav>
+            <NavigationMenu />
           </div>
         </div>
